+++ schulcloud/File.swift
@@ -29,23 +29,6 @@ final class File: NSManagedObject {
     
 }
 
-//// MARK: Generated accessors for contents
-//extension File {
-//    
-//    @objc(addContentsObject:)
-//    @NSManaged public func addToContents(_ value: File)
-//    
-//    @objc(removeContentsObject:)
-//    @NSManaged public func removeFromContents(_ value: File)
-//    
-//    @objc(addContents:)
-//    @NSManaged public func addToContents(_ values: NSSet)
-//    
-//    @objc(removeContents:)
-//    @NSManaged public func removeFromContents(_ values: NSSet)
-//    
-//}
-
 extension File {
     
     struct Permissions : OptionSet {
