+++ schulcloud/Sync/Pullable.swift
@@ -29,20 +29,6 @@ extension Pullable where Self: NSManagedObject {
         return managedObject
     }
 
-//    fileprivate func findIncludedObject(for objectIdentifier: ResourceIdentifier, in includes: [ResourceData]?) -> ResourceData? {
-//        guard let includedData = includes else {
-//            return nil
-//        }
-//
-//        return includedData.first { item in
-//            guard let identifier = try? ResourceIdentifier(object: item) else {
-//                return false
-//            }
-//            return objectIdentifier.id == identifier.id && objectIdentifier.type == identifier.type
-//        }
-//    }
-
-
     func updateRelationship<A>(forKeyPath keyPath: ReferenceWritableKeyPath<Self, A>,
                                forKey key: KeyType,
                                fromObject object: ResourceData,
@@ -58,19 +44,6 @@ extension Pullable where Self: NSManagedObject {
         default:
             throw SynchronizationError.missingIncludedResource(from: Self.self, to: A.self, withKey: key)
         }
-
-//        let resourceIdentifier = try object.value(for: "\(key).data") as ResourceIdentifier
-//
-//        if let includedObject = self.findIncludedObject(for: resourceIdentifier, in: includes) {
-//            var existingObject = self[keyPath: keyPath] // TODO: also check if id is equal. update() does not updates the id
-//            do {
-//                try existingObject.update(withObject: includedObject, withAdditionalSyncData: additionalSyncData, inContext: context)
-//            } catch let error as MarshalError {
-//                throw NestedMarshalError.nestedMarshalError(error, includeType: A.type, includeKey: key)
-//            }
-//        } else {
-//            throw SynchronizationError.missingIncludedResource(from: Self.self, to: A.self, withKey: key)
-//        }
     }
 
     func updateRelationship<A>(forKeyPath keyPath: ReferenceWritableKeyPath<Self, A?>,
@@ -103,36 +76,6 @@ extension Pullable where Self: NSManagedObject {
             // relationship does not exist, so we reset delete the possible relationship
             self[keyPath: keyPath] = nil
         }
-
-//
-//        guard let resourceIdentifier = try? object.value(for: "\(key).data") as ResourceIdentifier else {
-//            // relationship does not exist, so we reset delete the possible relationship
-//            self[keyPath: keyPath] = nil
-//            return
-//        }
-//
-//        if let includedObject = self.findIncludedObject(for: resourceIdentifier, in: includes) {
-//            do {
-//                if var existingObject = self[keyPath: keyPath] { // TODO: also check if id is equal. update() does not updates the id
-//                    try existingObject.update(withObject: includedObject, withAdditionalSyncData: additionalSyncData, inContext: context)
-//                } else {
-//                    if var fetchedResource = try SyncEngine.findExistingResource(withId: resourceIdentifier.id, ofType: A.self, inContext: context) {
-//                        try fetchedResource.update(withObject: includedObject, withAdditionalSyncData: additionalSyncData, inContext: context)
-//                        self[keyPath: keyPath] = fetchedResource
-//                    } else {
-//                        self[keyPath: keyPath] = try A.value(from: includedObject, withAdditionalSyncData: additionalSyncData, inContext: context)
-//                    }
-//                }
-//            } catch let error as MarshalError {
-//                throw NestedMarshalError.nestedMarshalError(error, includeType: A.type, includeKey: key)
-//            }
-//        } else {
-//            if let fetchedResource = try SyncEngine.findExistingResource(withId: resourceIdentifier.id, ofType: A.self, inContext: context) {
-//                self[keyPath: keyPath] = fetchedResource
-//            } else {
-//                log.info("relationship update saved (\(Self.type) --> \(A.type)?)")
-//            }
-//        }
     }
 
     func updateRelationship<A>(forKeyPath keyPath: ReferenceWritableKeyPath<Self, Set<A>>,
@@ -140,10 +83,6 @@ extension Pullable where Self: NSManagedObject {
                                fromObject object: ResourceData,
                                with context: SynchronizationContext) throws where A: NSManagedObject & Pullable {
         var currentObjects = Set(self[keyPath: keyPath])
-
-
-
-        //        let resourceIdentifiers = try object.value(for: "\(key).data") as [ResourceIdentifier]
 
         do {
             switch context.strategy.findIncludedObjects(forKey: key, ofObject: object, with: context) {
@@ -179,33 +118,6 @@ extension Pullable where Self: NSManagedObject {
             case .notExisting:
                 break
             }
-//            for resourceIdentifier in resourceIdentifiers {
-//                if var currentObject = currentObjects.first(where: { $0.id == resourceIdentifier.id }) {
-//                    if let includedObject = self.findIncludedObject(for: resourceIdentifier, in: includes) {
-//                        try currentObject.update(withObject: includedObject, withAdditionalSyncData: additionalSyncData, inContext: context)
-//                    }
-//
-//                    if let index = currentObjects.index(where: { $0 == currentObject }) {
-//                        currentObjects.remove(at: index)
-//                    }
-//                } else {
-//                    if let includedObject = self.findIncludedObject(for: resourceIdentifier, in: includes) {
-//                        if var fetchedResource = try SyncEngine.findExistingResource(withId: resourceIdentifier.id, ofType: A.self, inContext: context) {
-//                            try fetchedResource.update(withObject: includedObject, withAdditionalSyncData: additionalSyncData, inContext: context)
-//                            self[keyPath: keyPath].insert(fetchedResource)
-//                        } else {
-//                            let newObject = try A.value(from: includedObject, withAdditionalSyncData: additionalSyncData, inContext: context)
-//                            self[keyPath: keyPath].insert(newObject)
-//                        }
-//                    } else {
-//                        if let fetchedResource = try SyncEngine.findExistingResource(withId: resourceIdentifier.id, ofType: A.self, inContext: context) {
-//                            self[keyPath: keyPath].insert(fetchedResource)
-//                        } else {
-//                            log.info("relationship update saved (\(Self.type) --> Set<\(A.type)>)")
-//                        }
-//                    }
-//                }
-//            }
         } catch let error as MarshalError {
             throw NestedMarshalError.nestedMarshalError(error, includeType: A.type, includeKey: key)
         }
@@ -232,6 +144,7 @@ extension Pullable where Self: NSManagedObject {
 }
 
 class AbstractPullableContainer<A, B> where A: NSManagedObject & Pullable, B: NSManagedObject & AbstractPullable {
+
     let resource: A
     let keyPath: ReferenceWritableKeyPath<A, B?>
     let key: KeyType
@@ -267,22 +180,9 @@ class AbstractPullableContainer<A, B> where A: NSManagedObject & Pullable, B: NS
                 throw NestedMarshalError.nestedMarshalError(error, includeType: C.type, includeKey: key)
             }
         default:
-            // TODO throw error
+            // TODO throw error?
             break
         }
-
-
-//        if let includedObject = self.resource.findIncludedObject(for: resourceIdentifier, in: self.includes) {
-//            do {
-//                if var existingObject = self.resource[keyPath: self.keyPath] as? C{
-//                    try existingObject.update(withObject: includedObject, withAdditionalSyncData: self.additionalSyncData, inContext: context)
-//                } else if let newObject = try C.value(from: includedObject, withAdditionalSyncData: self.additionalSyncData, inContext: context) as? B {
-//                    self.resource[keyPath: self.keyPath] = newObject
-//                }
-//            } catch let error as MarshalError {
-//                throw NestedMarshalError.nestedMarshalError(error, includeType: C.type, includeKey: key)
-//            }
-//        }
     }
 
 }
