--- schulcloud/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  schulcloud
-//
-//  Created by Carl Gödecken on 05.05.17.
-//  Copyright © 2017 Hasso-Plattner-Institut. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UITableViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
